# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/fleet/{platform}/imports": {
      "post": {
        "operationId": "ImportController_upload",
        "parameters": [
          {
            "name": "platform",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "gojek",
                "grab"
              ],
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": [
                  "file",
                  "month",
                  "year"
                ],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  },
                  "month": {
                    "type": "integer",
                    "example": 7
                  },
                  "year": {
                    "type": "integer",
                    "example": 2026
                  }
                }
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Upload a CSV/XLSX for a period — parsed asynchronously",
        "tags": [
          "admin-fleet-imports"
        ]
      },
      "get": {
        "operationId": "ImportController_list",
        "parameters": [
          {
            "name": "platform",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "gojek",
                "grab"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "List import batches",
        "tags": [
          "admin-fleet-imports"
        ]
      }
    },
    "/admin/fleet/{platform}/imports/{id}": {
      "get": {
        "operationId": "ImportController_getById",
        "parameters": [
          {
            "name": "platform",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "gojek",
                "grab"
              ],
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Import status/progress (also streamed via Socket.IO /rt)",
        "tags": [
          "admin-fleet-imports"
        ]
      },
      "delete": {
        "operationId": "ImportController_rollback",
        "parameters": [
          {
            "name": "platform",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "gojek",
                "grab"
              ],
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "202": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Rollback a whole import batch (queued)",
        "tags": [
          "admin-fleet-imports"
        ]
      }
    },
...
    "/admin/fleet/{platform}/targets/{plate}": {
      "get": {
        "operationId": "TargetsController_get",
        "parameters": [
          {
            "name": "platform",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "gojek",
                "grab"
              ],
              "type": "string"
            }
          },
          {
            "name": "plate",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Read driver/target metadata for a plate",
        "tags": [
          "admin-fleet-targets"
        ]
      },
      "put": {
        "operationId": "TargetsController_upsert",
        "parameters": [
          {
            "name": "platform",
            "required": true,
            "in": "path",
            "schema": {
              "enum": [
                "gojek",
                "grab"
              ],
              "type": "string"
            }
          },
          {
            "name": "plate",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertTargetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "Create/update driver/target metadata (upsert by plate)",
        "tags": [
          "admin-fleet-targets"
        ]
      }
    },
...
    "/partner/portal/orders/{id}": {
      "get": {
        "operationId": "PortalController_order",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "session": []
          }
        ],
        "summary": "One own order detail (403 for another partner’s order)",
        "tags": [
          "partner-portal"
        ]
      }
    },
...
    "/partner/v1/orders": {
      "post": {
        "operationId": "PartnerOrdersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePartnerOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "partner-api-key": []
          }
        ],
        "summary": "Create an order (replaces legacy GET /partner/v1/order/create)",
        "tags": [
          "partner-v1"
        ]
      },
      "get": {
        "operationId": "PartnerOrdersController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "example": 1,
              "type": "string"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "example": 50,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "partner-api-key": []
          }
        ],
        "summary": "Order history for the authenticated partner",
        "tags": [
          "partner-v1"
        ]
      }
    },
    "/partner/v1/orders/{id}": {
      "get": {
        "operationId": "PartnerOrdersController_detail",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "partner-api-key": []
          }
        ],
        "summary": "One order's detail (own orders only — 404 for any other order)",
        "tags": [
          "partner-v1"
        ]
      }
    }
  },
  "info": {
    "title": "fleet-taxi-dashboard-api",
    "description": "Fleet/deposit-reconciliation dashboard API. All money values are integer rupiah serialized as JSON numbers. Every response uses the standard envelope: `{ success: true, data, meta? }` on success, `{ success: false, error: { code, message, details? } }` on error.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sid"
      },
      "partner-api-key": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Partner API key (external /partner/v1 only)"
      }
    },
    "schemas": {
...
      "UpsertTargetDto": {
        "type": "object",
        "properties": {
          "fleetTarget": {
            "type": "number",
            "description": "Daily target in integer rupiah (Gojek)"
          },
          "rentalPartner": {
            "type": "string"
          },
          "deliveryBatch": {
            "type": "string",
            "description": "Gojek only"
          },
          "serviceArea": {
            "type": "string",
            "description": "Gojek only"
          },
          "vehicleType": {
            "type": "string"
          },
          "regionId": {
            "type": "number",
            "description": "Gojek only"
          },
          "city": {
            "type": "string",
            "description": "Grab only"
          }
        }
      },
      "CreatePartnerOrderDto": {
        "type": "object",
        "properties": {
          "pickupCode": {
            "type": "string",
            "example": "BHISA_CAWANG",
            "description": "Pool code (whitelisted)"
          },
          "destinationCode": {
            "type": "string",
            "example": "EVISTA_HALIM"
          },
          "carTypesId": {
            "type": "number",
            "example": 1
          },
          "pickupAt": {
            "type": "string",
            "example": "2026-07-10 09:00:00",
            "description": "ISO 8601, or \"YYYY-MM-DD HH:mm:ss\" interpreted as Asia/Jakarta"
          },
          "passengerDetails": {
            "type": "object",
            "description": "Free-form passenger details object (stored as JSON; size-bounded)"
          }
        },
        "required": [
          "pickupCode",
          "destinationCode",
          "carTypesId",
          "pickupAt"
        ]
      }
    }
  }
}